project/requirements.txt
# Core libraries
numpy>=1.19.0
opencv-python>=4.5.4
PyYAML>=6.0
matplotlib>=3.5.1
pillow>=9.0.0
tqdm>=4.64.0

# Deep learning
torch>=1.12.0
torchvision>=0.13.0
ultralytics>=8.0.0  # YOLOv8

# Tracking
scikit-learn>=1.0.2
scipy>=1.8.0
lap>=0.4.0  # Hungarian algorithm for DeepSORT

# Visualization and UI
streamlit>=1.10.0
plotly>=5.9.0
pandas>=1.4.2

# Utilities
psutil>=5.9.0 